+++ inc/extras.php
@@ -133,7 +133,7 @@ add_filter( 'onepress_the_content', 'convert_smilies' );
 
 if ( ! function_exists( 'onepress_is_wc_active' ) ) {
     function onepress_is_wc_active(){
-        if ( function_exists( 'is_woocommerce' ) ) {
+        if ( class_exists( 'WooCommerce' ) || function_exists( 'is_woocommerce' ) ) {
            return true;
         }
         return false;
@@ -325,5 +325,70 @@ if ( ! function_exists( 'onepress_after_section' ) ) {
 
 add_action( 'onepress_before_section_part', 'onepress_before_section', 10, 2 );
 add_action( 'onepress_after_section_part', 'onepress_after_section', 10, 2 );
+
+
+if ( onepress_is_wc_active() ) {
+    /**
+     * Template pages
+    */
+
+    if ( ! function_exists( 'woocommerce_content' ) ) {
+
+        /**
+         * Output WooCommerce content.
+         *
+         * This function is only used in the optional 'woocommerce.php' template.
+         * which people can add to their themes to add basic woocommerce support.
+         * without hooks or modifying core templates.
+         * @since 2.0.6
+         */
+        function woocommerce_content() {
+
+            if ( is_singular( 'product' ) ) {
+
+                while ( have_posts() ) :
+                    the_post();
+                    wc_get_template_part( 'content', 'single-product' );
+                endwhile;
+
+            } else {
+
+                ?>
+                <?php if ( apply_filters( 'woocommerce_show_page_title', true ) ) : ?>
+                <div class="entry-header">
+                    <h1 class="page-title entry-title"><?php woocommerce_page_title(); ?></h1>
+                </div>
+                <?php endif; ?>
+
+                <?php do_action( 'woocommerce_archive_description' ); ?>
+
+                <?php if ( have_posts() ) : ?>
+
+                    <?php do_action( 'woocommerce_before_shop_loop' ); ?>
+
+                    <?php woocommerce_product_loop_start(); ?>
+
+                    <?php if ( wc_get_loop_prop( 'total' ) ) : ?>
+                        <?php while ( have_posts() ) : ?>
+                            <?php the_post(); ?>
+                            <?php wc_get_template_part( 'content', 'product' ); ?>
+                        <?php endwhile; ?>
+                    <?php endif; ?>
+
+                    <?php woocommerce_product_loop_end(); ?>
+
+                    <?php do_action( 'woocommerce_after_shop_loop' ); ?>
+
+                <?php else : ?>
+
+                    <?php do_action( 'woocommerce_no_products_found' ); ?>
+
+                <?php
+                endif;
+
+            }
+        }
+    }
+}