+++ changelog.md
@@ -3,6 +3,9 @@ CHANGELOG
 ====================================================================
 ## 2.0.6
 * FIXED: JS error issue.
+* FIXED: Duplicate title on single post.
+* FIXED: Duplicate title on product archive.
+
 
 ## 2.0.5
 * NEW: Support PirateForms.